refactor(activerecord): db_warnings log arm logs through Base.logger, one check_validity! in Association#initialize, two-call validates_uniqueness_of (RFC 0112, RFC 0122)
- `ActiveRecord.dbWarningsAction`'s `"log"` arm drops the `console.warn`
  fallback and interpolates the warning's own class, mirroring
  active_record.rb:240-245. The PG warning cases wrap their bodies in
  `withDbWarningsAction` (test_case.rb:164) and the log cases — PG and
  mysql2 — assert on a `Base.logger` double, as Rails does
  (postgresql_adapter_test.rb:630-637, warnings_test.rb:29-37).
- `Association#initialize` makes the single `reflection.check_validity!`
  call Rails makes (association.rb:35-45); `checkKlass` is gone, the
  unknown-class NameError now surfacing from `check_validity!`'s own
  `klass` access (reflection.rb:618-628).
- `validatesUniquenessOf` is `_merge_attributes` + `validates_with`
  (uniqueness.rb:291-293); the `:scope` check lives only in
  `UniquenessValidator#initialize`, at Rails' site (uniqueness.rb:11-14).
- Relocates the trails-only tail of `to-sql.test.ts` (Quoted/Casted
  collapse, Lock, OuterJoin guard, Node-named Table, the Update/Delete
  manager subselects, and the whole `ArelQuoter / defaultQuoter wiring`
  describe) to `to-sql.trails.test.ts` verbatim.
